+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add base_url argument to Bucket.create_query_link, [PR-142](https://github.com/reductstore/reduct-py/pull/142)
+
 ## [1.17.1] - 2025-11-11
 
 ### Fixed